Refactor create position to hook
